+++ monew/src/main/java/com/codeit/team2/monew/module/domain/comment/exception/CommentPermissionDeniedException.java
@@ -1,7 +1,6 @@
 package com.codeit.team2.monew.module.domain.comment.exception;
 
 import com.codeit.team2.monew.module.common.exception.BaseException;
-import com.codeit.team2.monew.module.domain.comment.code.CommentErrorCode;
 import java.util.Map;
 import java.util.UUID;
