+++ monew/src/main/java/com/codeit/team2/monew/module/domain/interest/exception/InvalidSimilarityInputException.java
@@ -1,7 +1,6 @@
 package com.codeit.team2.monew.module.domain.interest.exception;
 
 import com.codeit.team2.monew.module.common.exception.BaseException;
-import com.codeit.team2.monew.module.domain.interest.code.InterestErrorCode;
 
 public class InvalidSimilarityInputException extends BaseException {
