+++ monew/src/main/java/com/codeit/team2/monew/module/domain/subscription/exception/DuplicateSubscriptionException.java
@@ -1,7 +1,6 @@
 package com.codeit.team2.monew.module.domain.subscription.exception;
 
 import com.codeit.team2.monew.module.common.exception.BaseException;
-import com.codeit.team2.monew.module.domain.subscription.code.SubscriptionErrorCode;
 import java.util.Map;
 import java.util.UUID;
